+++ CHANGELOG.md
@@ -7,6 +7,8 @@ All notable changes to this project will be documented in this file.
 - new datetime alias (WIP)
 
 ### Updates
+- better caret handling in colormask
+- change notation of optionalmarker, quantifiermarker, groupmarker
 - drop prevalidator and cardinality support in definitions
 - better determination of the NdxInitializer (alternations)
 - disable autocorrect on safari when disablePredictiveText is used